+++ tests/test_runner.adb
@@ -8,6 +8,7 @@ with Podmander.Agent.Host_Command_Tests;
 with Podmander.Agent.Host_Command.Result_Mapping_Tests;
 with Podmander.Controller.Agent.Repository_Tests;
 with Podmander.Controller.Registrar_Tests;
+with Podmander.Controller.Scheduler_Tests;
 with Podmander.Controller.Service.Repository_Tests;
 with Podmander.Controller.Service_Catalog.Repository_Tests;
 with Podmander.Controller_Tests;
@@ -33,6 +34,8 @@ procedure Test_Runner is
         (Result, Podmander.Controller.Agent.Repository_Tests.Suite);
       AUnit.Test_Suites.Add_Test
         (Result, Podmander.Controller.Registrar_Tests.Suite);
+      AUnit.Test_Suites.Add_Test
+        (Result, Podmander.Controller.Scheduler_Tests.Suite);
       AUnit.Test_Suites.Add_Test
         (Result, Podmander.Controller.Service.Repository_Tests.Suite);
       AUnit.Test_Suites.Add_Test